[Northamptonshire] use Alloy vector tiles for assets
Replace the code that uses the API with Vector tiles which don't require
and API key or paging. This code is still a bit work in progress as it
makes a few assumptions:
 * the zoom level is hard coded
 * adding assets to layers assumes things are not async when they are
 * we are using itemId which is not the correct thing

This also doesn't have anything for putting labels on assets because
that information doesn't seem to be in the tiles.

A final point is that the tiles can contain multiple assets at the same
coőrdinates and we don't do anything about that at the moment so the
user will just get the last one added when they click.