[Northamptonshire] improve alloy asset layer loading
This uses an event to notify that loading a request for a new set of
tiles has started and then uses counting to check we've fetched them all
and hence can get the layer to check for an asset.

It also uses start times to throw away requests for tiles in a previous
request if they come in after a new request has started.

This should fix issues with asset ids not added to reports.